// src/shared/api/generated/model/sessionRecordsResponseDtoRecordsItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Yoclick API
 * API multi-centro de reservas en marca blanca. Contrato canónico para la app.
 * OpenAPI spec version: 0.1.0
 */
import type { SessionRecordsResponseDtoRecordsItemBooking } from './sessionRecordsResponseDtoRecordsItemBooking.ts';
import type { SessionRecordsResponseDtoRecordsItemClient } from './sessionRecordsResponseDtoRecordsItemClient.ts';
import type { SessionRecordsResponseDtoRecordsItemStaff } from './sessionRecordsResponseDtoRecordsItemStaff.ts';

export type SessionRecordsResponseDtoRecordsItem = {
  booking: SessionRecordsResponseDtoRecordsItemBooking;
  client: SessionRecordsResponseDtoRecordsItemClient;
  staff: SessionRecordsResponseDtoRecordsItemStaff;
  /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     */
  plannedDurationSeconds: number;
  /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     * @nullable
     */
  actualDurationSeconds: number | null;
  isOpen: boolean;
  /** @nullable */
  notes: string | null;
};